Move tkey/assert.h -> assert.h
